requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

# -i https://pypi.org/simple


asgiref==3.4.1
autopep8==1.5.7
django-crispy-forms==1.13.0
django-split-settings==1.1.0
django==3.2.7
pycodestyle==2.7.0
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.2
toml==0.10.2
